fix(lore-0046): address Copilot review on PR #149
- InvocationAppearanceRow.amount + InvocationAppearanceItem.amount: i64 → i32
  to match the actual schema (soroban_invocations_appearances.amount is
  INTEGER, not BIGINT) and the contracts module's existing convention.
  Events amount stays i64 (column is BIGINT).
- Drop redundant .to_owned() in push_contract_union_arm and
  push_source_predicate. Tighten lifetimes so &str binds directly,
  matching sister-module convention (assets/contracts queries).
- Register EventAppearanceItem and InvocationAppearanceItem in
  ApiDoc::components(schemas(...)) so the generated OpenAPI spec
  includes the nested DTOs and downstream codegen resolves their
  $ref'd schemas.

+++ crates/api/src/transactions/dto.rs
@@ -103,7 +103,7 @@ pub struct InvocationAppearanceItem {
     /// Root caller G-StrKey. Per ADR 0034 nested-call hierarchy is XDR-only.
     pub caller_account: Option<String>,
     pub ledger_sequence: i64,
-    pub amount: i64,
+    pub amount: i32,
     pub created_at: DateTime<Utc>,
 }
